otsu and triangle get processed at the same time, and batch run mitometer
